// src/lib/api/generated/model/formQuestion.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Django Continuous Integration API
 * API documentation for the Django Continuous Integration project with tenant-aware access rules.
 * OpenAPI spec version: 1.0.0
 */
import type { QuestionTypeEnum } from './questionTypeEnum';

export interface FormQuestion {
  readonly id: string;
  /** @nullable */
  readonly tenant: string | null;
  readonly form: string;
  /** @maxLength 255 */
  label: string;
  question_type: QuestionTypeEnum;
  is_required?: boolean;
  readonly order: number;
  indicators?: string[];
  readonly created_at: string;
  readonly updated_at: string;
}